data: Move file review selection validation
File-review state persists review identity and coordinates pathless action
safety.

Selection containment checks also lived there, so status reporting and
batch review selection code imported validation state for pure selection
rules.

This commit addresses that by moving shown-selection filtering and scoped
line validation into data.file_review_selection_validation and routing
state/status/data callers through that boundary.

The state module keeps persistence and action-scope orchestration while
selection rules live beside the other file-review data helpers.
