+++ src/git_stage_batch/data/file_review/state.py
@@ -4,12 +4,10 @@ from __future__ import annotations
 
 import json
 import shlex
-from collections.abc import Iterable
-from dataclasses import asdict, dataclass
+from dataclasses import asdict
 
 from ...core.actionable_changes import ActionableSelectionReason
-from ...core.line_selection import LineRanges, LineSelection, parse_line_selection_ranges
-from ...batch import file_display
+from ...core.line_selection import parse_line_selection_ranges
 from ...exceptions import CommandError
 from ...i18n import _
 from ...utils.file_io import read_text_file_contents, write_text_file_contents
@@ -25,6 +23,10 @@ from .freshness import (
     review_state_matches_action as _review_state_matches_action,
     selected_change_kind_matches_review_source as _selected_change_kind_matches_review_source,
     selected_change_matches_review_state as _selected_change_matches_review_state,
+)
+from .selection_validation import (
+    shown_review_selections_for_action as _shown_review_selections_for_action,
+    validate_review_scoped_line_selection as _validate_review_scoped_line_selection,
 )
 from ..selected_change.store import SelectedChangeKind, read_selected_change_kind
 
@@ -223,41 +225,6 @@ def _format_pages(pages: set[int]) -> str:
     return format_line_ids(sorted(pages))
 
 
-def _format_line_ranges(selection: LineRanges) -> str:
-    return selection.to_line_spec()
-
-
-def _coerce_line_ranges(selection: LineSelection | Iterable[int]) -> LineRanges:
-    if isinstance(selection, LineRanges):
-        return selection
-    return LineRanges.from_lines(selection)
-
-
-@dataclass(frozen=True)
-class _ReviewValidationGroup:
-    display_ids: LineRanges
-    is_splittable: bool
-
-
-def shown_review_selections_for_action(
-    review_state: _records.FileReviewState,
-    action: _records.FileReviewAction | str,
-) -> list[_records.FileReviewSelectionState]:
-    """Return actionable selections fully contained by the shown review pages."""
-    review_action = _coerce_review_action(action)
-    shown_pages = (
-        set(range(1, review_state.page_count + 1))
-        if review_state.entire_file_shown else
-        set(review_state.shown_pages)
-    )
-    return [
-        selection
-        for selection in review_state.selections
-        if review_action in selection.actions
-        and set(range(selection.first_page, selection.last_page + 1)).issubset(shown_pages)
-    ]
-
-
 def fresh_batch_review_selections_for_action(
     batch_name: str,
     file_path: str,
@@ -289,7 +256,7 @@ def fresh_batch_review_selections_for_action(
             )
         )
 
-    return shown_review_selections_for_action(review_state, action)
+    return _shown_review_selections_for_action(review_state, action)
 
 
 def _print_stale_or_mismatched_file_review_help(action: str, review_state: _records.FileReviewState) -> None:
@@ -612,74 +579,6 @@ def resolve_live_line_action_scope(
     return _records.ActionScopeResolution(file=file, review_state=review_state)
 
 
-def validate_review_scoped_line_selection(
-    requested_ids: LineSelection | Iterable[int],
-    valid_selections: Iterable[_records.FileReviewSelectionState],
-) -> None:
-    """Validate a union of complete actionable review selections."""
-    requested_ranges = _coerce_line_ranges(requested_ids)
-    groups: list[_ReviewValidationGroup] = []
-    for selection in valid_selections:
-        display_ids = LineRanges.from_lines(selection.display_ids)
-        if display_ids:
-            groups.append(
-                _ReviewValidationGroup(
-                    display_ids=display_ids,
-                    is_splittable=selection.is_splittable,
-                )
-            )
-
-    def can_cover(remaining_ids: LineRanges) -> bool:
-        if not remaining_ids:
-            return True
-        first_id = remaining_ids.first()
-        if first_id is None:
-            return True
-        for group in groups:
-            if first_id not in group.display_ids:
-                continue
-            if group.is_splittable:
-                selected_from_group = remaining_ids.intersection(group.display_ids)
-                if can_cover(remaining_ids.difference(selected_from_group)):
-                    return True
-                continue
-            if group.display_ids.difference(remaining_ids):
-                continue
-            if can_cover(remaining_ids.difference(group.display_ids)):
-                return True
-        return False
-
-    if can_cover(requested_ranges):
-        return
-
-    matched_ids = LineRanges.empty()
-    for group in groups:
-        if group.is_splittable:
-            matched_ids = matched_ids.union(requested_ranges.intersection(group.display_ids))
-        elif not group.display_ids.difference(requested_ranges):
-            matched_ids = matched_ids.union(group.display_ids)
-
-    outside_ids = requested_ranges.difference(matched_ids)
-    for group in groups:
-        if group.is_splittable:
-            continue
-        overlap = outside_ids.intersection(group.display_ids)
-        if overlap and overlap != group.display_ids:
-            raise CommandError(
-                _("Line selection #{requested} only partly selects a reviewed change.\nUse: --line {required}").format(
-                    requested=_format_line_ranges(requested_ranges),
-                    required=_format_line_ranges(group.display_ids),
-                )
-            )
-
-    if outside_ids:
-        raise CommandError(
-            _("Line selection #{ids} is not valid from the current file review.").format(
-                ids=_format_line_ranges(outside_ids),
-            )
-        )
-
-
 def validate_pathless_review_line_action(
     action: _records.FileReviewAction | str,
     line_id_specification: str,
@@ -734,6 +633,6 @@ def validate_pathless_review_line_action(
     except ValueError as error:
         raise CommandError(str(error)) from error
 
-    valid_selections = shown_review_selections_for_action(review_state, review_action)
-    validate_review_scoped_line_selection(requested_ids, valid_selections)
+    valid_selections = _shown_review_selections_for_action(review_state, review_action)
+    _validate_review_scoped_line_selection(requested_ids, valid_selections)
     return review_state
